+++ src/Http/Requests/CP/CreateRequest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace DoubleThreeDigital\Runway\Http\Requests;
+namespace DoubleThreeDigital\Runway\Http\Requests\CP;
 
 use Illuminate\Foundation\Http\FormRequest;
 use Statamic\Facades\User;
 
-class StoreRequest extends FormRequest
+class CreateRequest extends FormRequest
 {
     public function authorize()
     {
